+++ src/arstists/arstists.service.ts
@@ -8,27 +8,78 @@ export class ArtistsService {
   constructor(
     @InjectRepository(Artist)
     private usersRepository: Repository<Artist>,
-  ) {}
+  ) { }
 
+
+  //  new artist  
   async create(userData: Partial<Artist>): Promise<Artist> {
     const newUser = this.usersRepository.create(userData);
     return this.usersRepository.save(newUser);
   }
 
+
+  // get all
   async findAll(): Promise<Artist[]> {
     return this.usersRepository.find();
   }
 
-  async findOne(id: number): Promise<Artist> {
-    return this.usersRepository.findOneBy({ id });
+
+  // by id
+  async findOne(id: number): Promise<{ success: boolean; message: string; data?: Artist }> {
+    let [checkArtist] = await this.usersRepository.findBy({ id })
+    if (!checkArtist) {
+      return {
+        success: false,
+        message: 'Not found artist'
+      }
+    }
+    return {
+      success: true,
+      message: `Artist by id ${id}`,
+      data: checkArtist
+    }
   }
 
-  async update(id: number, userData: Partial<Artist>): Promise<Artist> {
-    await this.usersRepository.update(id, userData);
-    return this.findOne(id);
+
+  // update artist
+  async update(id: number, userData: Partial<Artist>): Promise<{ success: boolean; message: string; data?: Artist }> {
+    try {
+      let [checkArtist] = await this.usersRepository.findBy({ id })
+      if (!checkArtist) {
+        return {
+          success: false,
+          message: 'Not found artist'
+        }
+      }
+
+      await this.usersRepository.update(id, userData);
+      return {
+        success: true,
+        message: `Artist successfully updated`,
+        data: checkArtist
+      }
+    } catch (error: any) {
+      return error.message
+    }
   }
 
-  async remove(id: number): Promise<void> {
+
+  // remove artist
+  async remove(id: number): Promise<{ success: boolean; message: string; }> {
+    let [checkUser] = await this.usersRepository.findBy({
+      id
+    })
+
+    if (!checkUser) {
+      return {
+        success: false,
+        message: 'Not found artist'
+      }
+    }
     await this.usersRepository.delete(id);
+    return {
+      success: true,
+      message: `Successfully delete artist`,
+    }
   }
 }
